drif eoi moved to dedicated folder
