+++ drr/src/UI/src/app/drif/drif-eoi/drif-eoi-step-3/drif-eoi-step-3.component.ts
@@ -20,16 +20,16 @@ import {
   RxFormControl,
 } from '@rxweb/reactive-form-validators';
 import { distinctUntilChanged } from 'rxjs';
-import { FundingType } from '../../../model';
+import { FundingType } from '../../../../model';
 
-import { DrrCurrencyInputComponent } from '../../shared/controls/drr-currency-input/drr-currency-input.component';
-import { DrrInputComponent } from '../../shared/controls/drr-input/drr-input.component';
-import { DrrSelectComponent } from '../../shared/controls/drr-select/drr-select.component';
-import { DrrTextareaComponent } from '../../shared/controls/drr-textarea/drr-textarea.component';
+import { DrrCurrencyInputComponent } from '../../../shared/controls/drr-currency-input/drr-currency-input.component';
+import { DrrInputComponent } from '../../../shared/controls/drr-input/drr-input.component';
+import { DrrSelectComponent } from '../../../shared/controls/drr-select/drr-select.component';
+import { DrrTextareaComponent } from '../../../shared/controls/drr-textarea/drr-textarea.component';
 import {
   FundingInformationForm,
   FundingInformationItemForm,
-} from '../drif-eoi/drif-eoi-form';
+} from '../drif-eoi-form';
 
 @Component({
   selector: 'drif-eoi-step-3',
