+++ drr/src/UI/src/app/drif/drif-eoi/drif-eoi-step-4/drif-eoi-step-4.component.ts
@@ -13,8 +13,8 @@ import { MatInputModule } from '@angular/material/input';
 import { MatRadioModule } from '@angular/material/radio';
 import { TranslocoModule } from '@ngneat/transloco';
 import { IFormGroup } from '@rxweb/reactive-form-validators';
-import { DrrTextareaComponent } from '../../shared/controls/drr-textarea/drr-textarea.component';
-import { LocationInformationForm } from '../drif-eoi/drif-eoi-form';
+import { DrrTextareaComponent } from '../../../shared/controls/drr-textarea/drr-textarea.component';
+import { LocationInformationForm } from '../drif-eoi-form';
 
 @Component({
   selector: 'drif-eoi-step-4',
