+++ drr/src/UI/src/app/drif/drif-eoi/drif-eoi-view/drif-eoi-view.component.html
